+++ docs/components/example/check-select-box-react-hook-form.tsx
@@ -58,7 +58,7 @@ export default function CheckSelectBoxReactHookForm() {
         <ActionButton type="button" variant="neutralWeak" onClick={() => setValue("mango", true)}>
           mango 선택
         </ActionButton>
-        <ActionButton type="submit" flexGrow>
+        <ActionButton type="submit" flexGrow={1}>
           제출
         </ActionButton>
       </HStack>