+++ docs/components/mdx-components.ts
@@ -14,6 +14,7 @@ import { TypeTable } from "fumadocs-ui/components/type-table";
 import defaultMdxComponents from "fumadocs-ui/mdx";
 import { AtomIcon } from "lucide-react";
 import { MDXComponents } from "mdx/types";
+import { BreezeManualInstallation } from "./breeze-manual-installation";
 import { IconLibrary } from "./iconography/icons";
 import { ColorMigrationIndex } from "./migration/color-migration-index";
 import { V2Icon, V2IconColor, V3Icon } from "./migration/icon";
@@ -30,6 +31,7 @@ export const mdxComponents: MDXComponents = {
   ComponentExample,
   TokenReference,
   ComponentSpecBlock,
+  BreezeManualInstallation,
   Tab,
   Tabs,
   Step,