+++ examples/stackflow-spa/src/seed-design/ui/extended-action-sheet.tsx
@@ -2,6 +2,7 @@
 
 import { ExtendedActionSheet as SeedExtendedActionSheet } from "@seed-design/react";
 import { forwardRef } from "react";
+import type * as React from "react";
 
 /**
  * @deprecated Use menu-sheet instead.