+++ packages/react/src/components/Divider/Divider.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Box, type BoxProps } from "../Box/Box";
 
-export interface DividerProps {
+export interface DividerProps extends Omit<React.HTMLAttributes<HTMLHRElement>, "color"> {
   /**
    * The HTML element to use for the divider.
    * Keep in mind that "hr" elements are read by screen readers as a semantic break with an implicit role="separator"
@@ -18,43 +18,44 @@ export interface DividerProps {
   /**
    * @default 1
    */
-  thickness?: BoxProps["borderBottomWidth"];
+  thickness?: BoxProps["borderWidth"];
 
   /**
    * @default "horizontal"
    */
   orientation?: "horizontal" | "vertical";
-
-  role?: Extract<React.AriaRole, "separator">;
 }
 
-export const Divider = React.forwardRef<HTMLHRElement, DividerProps>((props, ref) => {
-  const {
-    as = "hr",
-    color = "stroke.neutralMuted",
-    thickness = 1,
-    orientation = "horizontal",
-    role,
-    ...rest
-  } = props;
-
-  return (
-    <Box
-      ref={ref}
-      as={as}
-      role={role}
-      // if hr, aria-orientation=horizontal is implied if not specified
-      // if not hr, aria-orientation is needed
-      {...(((as === "hr" && orientation !== "horizontal") ||
-        (as !== "hr" && role === "separator")) && {
-        "aria-orientation": orientation,
-      })}
-      display="block"
-      borderColor={color}
-      borderWidth={0}
-      {...(orientation === "vertical" && { borderRightWidth: thickness })}
-      {...(orientation === "horizontal" && { borderBottomWidth: thickness })}
-      {...rest}
-    />
-  );
-});
+export const Divider = React.forwardRef<HTMLHRElement, DividerProps>(
+  (
+    {
+      as = "hr",
+      color = "stroke.neutralMuted",
+      thickness = 1,
+      orientation = "horizontal",
+      ...props
+    },
+    ref,
+  ) => {
+    return (
+      <Box
+        ref={ref}
+        as={as}
+        // if hr, aria-orientation=horizontal is implied if not specified
+        {...(((as === "hr" &&
+          orientation !== "horizontal" &&
+          (props.role === undefined || props.role === "separator")) ||
+          // if not hr but role is separator aria-orientation is needed
+          (as !== "hr" && props.role === "separator")) && {
+          "aria-orientation": orientation,
+        })}
+        display="block"
+        borderColor={color}
+        borderWidth={0}
+        {...(orientation === "vertical" && { borderRightWidth: thickness })}
+        {...(orientation === "horizontal" && { borderBottomWidth: thickness })}
+        {...props}
+      />
+    );
+  },
+);