+++ packages/react/src/components/List/List.namespace.ts
@@ -1,0 +1,16 @@
+export {
+  ListContent as Content,
+  ListDetail as Detail,
+  ListItem as Item,
+  ListPrefix as Prefix,
+  ListRoot as Root,
+  ListSuffix as Suffix,
+  ListTitle as Title,
+  type ListContentProps as ContentProps,
+  type ListDetailProps as DetailProps,
+  type ListItemProps as ItemProps,
+  type ListPrefixProps as PrefixProps,
+  type ListRootProps as RootProps,
+  type ListSuffixProps as SuffixProps,
+  type ListTitleProps as TitleProps,
+} from "./List";