+++ packages/react/src/components/List/index.ts
@@ -1,0 +1,18 @@
+export {
+  ListContent,
+  ListDetail,
+  ListItem,
+  ListPrefix,
+  ListRoot,
+  ListSuffix,
+  ListTitle,
+  type ListContentProps,
+  type ListDetailProps,
+  type ListItemProps,
+  type ListPrefixProps,
+  type ListRootProps,
+  type ListSuffixProps,
+  type ListTitleProps,
+} from "./List";
+
+export * as List from "./List.namespace";